[Add] #16 - asset 추가 및 복약 계획 드롭메뉴 추가
